+++ packages/node/src/plugins/NodeFederationPlugin.ts
@@ -39,7 +39,7 @@ interface Context {
 //TODO: should use Template system like LoadFileChunk runtime does.
 //TODO: global.webpackChunkLoad could use a better convention? I have to use a special http client to get out of my infra firewall
 const executeLoadTemplate = `
-    function executeLoad(remoteUrl) {
+    function executeLoad(remoteUrl, retry) {
       function extractUrlAndGlobal(urlAndGlobal) {
         var index = urlAndGlobal.indexOf("@");
         if (index <= 0 || index === urlAndGlobal.length - 1) {
@@ -57,8 +57,7 @@ const executeLoadTemplate = `
             return res.text();
           }).then(function(scriptContent){
             try {
-              const vmContext = { exports, require, module, global, __filename, __dirname, URL };
-
+              const vmContext = { exports, require, module, global, __filename, __dirname, URL, ...global};
               const remote = vm.runInNewContext(scriptContent + '\\nmodule.exports', vmContext, { filename: 'node-federation-loader-' + moduleName + '.vm' });
 
               /* TODO: need something like a chunk loading queue, this can lead to async issues
@@ -76,14 +75,20 @@ const executeLoadTemplate = `
             reject(null)
           })
         }).catch((e)=>{
-        console.error('error',e);
+          console.error('error',e);
+
+          if(!retry) {
+            return executeLoad(remoteUrl, true);
+          }
           console.warn(moduleName,'is offline, returning fake remote')
           return {
             fake: true,
             get:(arg)=>{
               console.log('faking', arg,'module on', moduleName);
 
-              return ()=> Promise.resolve();
+              return Promise.resolve(()=>{
+              return ()=>null
+              });
             },
             init:()=>{}
           }
@@ -122,118 +127,43 @@ function buildRemotes(
     ${executeLoadTemplate}
     resolve(executeLoad(${JSON.stringify(config)}))
     }).then(remote=>{
-      console.log(remote);
-
-      return {
-      get: remote.get,
-      init: (args)=> {
-        console.log(args)
-        return remote.init(args)
+      if(!global.usedChunks || remote.fake) {
+        return remote;
       }
+
+
+   const proxy =  {
+      get: (arg)=>{
+        if(!global.__remote_scope__[${JSON.stringify(global)}].__initialized) {
+          try {proxy.init(__webpack_require__.S.default);} catch(e) {}
+        }
+        return remote.get(arg).then((f)=>{
+          const m = f();
+          return ()=>new Proxy(m, {
+            get: (target, prop)=>{
+            global.usedChunks.add(${JSON.stringify(global)} + "->" + arg);
+              return target[prop];
+            }
+          })
+        })
+      },
+        init: (args)=> {
+        global.__remote_scope__[${JSON.stringify(global)}].__initialized = true;
+          return remote.init(new Proxy(args, {
+            set: (target, prop, value)=>{
+              global.usedChunks.add(${JSON.stringify(global)} + "->" + prop);
+              target[prop] = value;
+              return true;
+            }
+          }))
+        }
     }
+
+    return proxy;
+
   });
   `;
       acc.buildTime[name] = loadTemplate;
-      return acc;
-    },
-    { runtime: {}, buildTime: {}, hot: {} } as {
-      runtime: EmptyObject;
-      buildTime: EmptyObject;
-      hot: EmptyObject;
-    }
-  );
-
-  //old design
-  return Object.entries(mfConf.remotes || {}).reduce(
-    (acc, [name, config]) => {
-      const hasMiddleware = config.startsWith('middleware ');
-      let middleware;
-      if (hasMiddleware) {
-        middleware = config.split('middleware ')[1];
-      } else {
-        middleware = `Promise.resolve(${JSON.stringify(config)})`;
-      }
-
-      const templateStart = `
-              var ${webpack.RuntimeGlobals.require} = ${
-        webpack.RuntimeGlobals.require
-      } ? ${
-        webpack.RuntimeGlobals.require
-      } : typeof arguments !== 'undefined' ? arguments[2] : false;
-               ${executeLoadTemplate}
-        global.loadedRemotes = global.loadedRemotes || {};
-        if (global.loadedRemotes[${JSON.stringify(name)}]) {
-          return global.loadedRemotes[${JSON.stringify(name)}]
-        }
-        // if using modern output, then there are no arguments on the parent function scope, thus we need to get it via a window global.
-
-      var shareScope = (${webpack.RuntimeGlobals.require} && ${
-        webpack.RuntimeGlobals.shareScopeMap
-      }) ? ${
-        webpack.RuntimeGlobals.shareScopeMap
-      } : global.__webpack_share_scopes__
-      var name = ${JSON.stringify(name)}
-      `;
-      const template = `(remotesConfig) => new Promise((res) => {
-      console.log('in template promise',JSON.stringify(remotesConfig))
-        executeLoad(remotesConfig).then((remote) => {
-
-          return Promise.resolve(remote.init(shareScope.default)).then(() => {
-            return remote
-          })
-        })
-          .then(function (remote) {
-            const proxy = {
-              get: remote.get,
-              chunkMap: remote.chunkMap,
-              path: remotesConfig.toString(),
-              init: (arg) => {
-                try {
-                  return remote.init(shareScope.default)
-                } catch (e) {
-                  console.log('remote container already initialized')
-                }
-              }
-            }
-            if (remote.fake) {
-              res(proxy);
-              return null
-            }
-
-            Object.assign(global.loadedRemotes, {
-              [name]: proxy
-            });
-
-            res(global.loadedRemotes[name])
-          })
-
-
-      })`;
-
-      acc.runtime[name] = `()=> ${middleware}.then((remoteConfig)=>{
-    console.log('remoteConfig runtime',remoteConfig);
-    if(!global.REMOTE_CONFIG) {
-        global.REMOTE_CONFIG = {};
-    }
-    global.REMOTE_CONFIG[${JSON.stringify(name)}] = remoteConfig;
-    ${templateStart}
-    const loadTemplate = ${template};
-    return loadTemplate(remoteConfig)
-    })`;
-
-      acc.buildTime[name] = `promise ${middleware}.then((remoteConfig)=>{
-            if(!global.REMOTE_CONFIG) {
-        global.REMOTE_CONFIG = {};
-    }
-    console.log('remoteConfig buildtime',remoteConfig);
-    global.REMOTE_CONFIG[${JSON.stringify(name)}] = remoteConfig;
-    ${templateStart};
-    const loadTemplate = ${template};
-    return loadTemplate(remoteConfig)
-    })`;
-
-      acc.hot[name] = `()=> ${middleware}`;
-
       return acc;
     },
     { runtime: {}, buildTime: {}, hot: {} } as {
@@ -257,6 +187,8 @@ class NodeFederationPlugin {
     this.context = context || ({} as Context);
     this.experiments = experiments || {};
   }
+
+
 
   apply(compiler: Compiler) {
     // When used with Next.js, context is needed to use Next.js webpack