+++ packages/utilities/src/utils/common.ts
@@ -9,10 +9,8 @@ import fs from 'fs';
 import fg from 'fast-glob';
 
 // @ts-expect-error - because process.env.REMOTES is not typed
-const remoteVars = (process.env.REMOTES || {}) as Record<
-  string,
-  Promise<any> | string | (() => Promise<any>)
->;
+const remoteVars = (process.env.REMOTES || {}) as Record<string,
+  Promise<any> | string | (() => Promise<any>)>;
 
 // split the @ syntax into url and global
 export const extractUrlAndGlobal = (urlAndGlobal: string): [string, string] => {
@@ -24,32 +22,32 @@ export const extractUrlAndGlobal = (urlAndGlobal: string): [string, string] => {
 };
 
 export const runtimeRemotes = Object.entries(remoteVars).reduce(function (
-  acc,
-  item
-) {
-  const [key, value] = item;
-  // if its an object with a thennable (eagerly executing function)
-  if (typeof value === 'object' && typeof value.then === 'function') {
-    acc[key] = { asyncContainer: value };
-  }
-  // if its a function that must be called (lazily executing function)
-  else if (typeof value === 'function') {
-    acc[key] = { asyncContainer: value() };
-  }
-  // if its just a string (global@url)
-  else if (typeof value === 'string') {
-    const [url, global] = extractUrlAndGlobal(value);
-    acc[key] = { global, url };
-  }
-  // we dont know or currently support this type
-  else {
-    // @ts-expect-error - because process.env.REMOTES is not typed
-    console.log('remotes process', process.env.REMOTES);
-    throw new Error(`[mf] Invalid value received for runtime_remote "${key}"`);
-  }
-  return acc;
-},
-{} as RuntimeRemotesMap);
+    acc,
+    item
+  ) {
+    const [key, value] = item;
+    // if its an object with a thennable (eagerly executing function)
+    if (typeof value === 'object' && typeof value.then === 'function') {
+      acc[key] = {asyncContainer: value};
+    }
+    // if its a function that must be called (lazily executing function)
+    else if (typeof value === 'function') {
+      acc[key] = {asyncContainer: value()};
+    }
+    // if its just a string (global@url)
+    else if (typeof value === 'string') {
+      const [url, global] = extractUrlAndGlobal(value);
+      acc[key] = {global, url};
+    }
+    // we dont know or currently support this type
+    else {
+      //@ts-ignore
+      console.log('remotes process', process.env.REMOTES);
+      throw new Error(`[mf] Invalid value received for runtime_remote "${key}"`);
+    }
+    return acc;
+  },
+  {} as RuntimeRemotesMap);
 
 export const remotes = runtimeRemotes;
 
@@ -92,7 +90,7 @@ export const injectScript = (
       function resolveRemoteGlobal() {
         const asyncContainer = window[
           remoteGlobal
-        ] as unknown as AsyncContainer;
+          ] as unknown as AsyncContainer;
         return resolve(asyncContainer);
       }